backend/requirements.txt
flask==3.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
redis==5.0.1
pytest==7.4.3
requests==2.31.0
